MOL-429: add Cypress Tests
