+++ .gitignore
@@ -3,3 +3,4 @@
 
 .reports
 .DS_Store
+/tests/Cypress/package-lock.json
