+++ tests/Cypress/.gitignore
@@ -1,0 +1,8 @@
+
+.DS_Store
+.idea
+/node_modules
+
+/cypress/videos
+/cypress/screenshots
+
