+++ tests/Cypress/Cypress.env.json
@@ -1,0 +1,2 @@
+{
+}
