+++ tests/Cypress/cypress/config/dev.json
@@ -1,0 +1,23 @@
+{
+    "chromeWebSecurity": false,
+    "watchForFileChanges": false,
+    "trashAssetsBeforeRuns": true,
+    "screenshotOnRunFailure": true,
+    "video": false,
+    "devices": [
+        {
+            "key": "desktop",
+            "name": "Desktop",
+            "width": 1920,
+            "height": 1080,
+            "userAgent": ""
+        },
+        {
+            "key": "ipad-landscape",
+            "name": "iPad (Landscape)",
+            "width": 1024,
+            "height": 768,
+            "userAgent": "Mozilla/5.0 (iPad; CPU OS 5_1 like Mac OS X) AppleWebKit/534.46 (KHTML, like Gecko) Version/5.1 Mobile/9B176 Safari/7534.48.3"
+        }
+    ]
+}
