+++ tests/Cypress/cypress/support/actions/storefront/account/LoginAction.js
@@ -1,0 +1,23 @@
+import LoginRepository from 'Repositories/storefront/account/LoginRepository';
+
+const repo = new LoginRepository();
+
+export default class LoginAction {
+
+    /**
+     *
+     * @param email
+     * @param password
+     */
+    doLogin(email, password) {
+
+        cy.visit('/account');
+
+        repo.getEmail().clear().type(email);
+        repo.getPassword().clear().type(password);
+
+        repo.getSubmitButton().click();
+    }
+
+}
+
