+++ tests/Cypress/cypress/support/actions/storefront/navigation/TopMenuAction.js
@@ -1,0 +1,15 @@
+import NavigationRepository from 'Repositories/storefront/navigation/NavigationRepository';
+
+export default class TopMenuAction {
+
+    /**
+     *
+     */
+    clickOnHome() {
+
+        const repo = new NavigationRepository();
+
+        repo.getHomeMenuItem().click();
+    }
+
+}
