+++ tests/Cypress/cypress/support/actions/storefront/products/ListingAction.js
@@ -1,0 +1,15 @@
+import ListingRepository from 'Repositories/storefront/products/ListingRepository';
+
+export default class ListingAction {
+
+    /**
+     *
+     */
+    clickOnFirstProduct() {
+
+        const repo = new ListingRepository();
+
+        repo.getFirstProduct().click();
+    }
+
+}
