+++ tests/Cypress/cypress/support/actions/storefront/products/PDPAction.js
@@ -1,0 +1,17 @@
+import PDPRepository from 'Repositories/storefront/products/PDPRepository';
+
+class PDPAction {
+
+    /**
+     *
+     */
+    addToCart() {
+
+        const repo = new PDPRepository();
+
+        repo.getAddToCartButton().click();
+    }
+
+}
+
+export default PDPAction;
