+++ tests/Cypress/cypress/support/repositories/storefront/account/LoginRepository.js
@@ -1,0 +1,27 @@
+export default class LoginRepository {
+
+    /**
+     *
+     * @returns {*}
+     */
+    getEmail() {
+        return cy.get('#loginMail');
+    }
+
+    /**
+     *
+     * @returns {*}
+     */
+    getPassword() {
+        return cy.get('#loginPassword');
+    }
+
+    /**
+     *
+     * @returns {*}
+     */
+    getSubmitButton() {
+        return cy.get('.login-submit > .btn')
+    }
+
+}
