+++ tests/Cypress/cypress/support/repositories/storefront/checkout/OffCanvasRepository.js
@@ -1,0 +1,11 @@
+export default class OffCanvasRepository {
+
+    /**
+     *
+     * @returns {*}
+     */
+    getCheckoutButton() {
+        return cy.get('.begin-checkout-btn');
+    }
+
+}
