+++ tests/Cypress/cypress/support/repositories/storefront/checkout/PaymentsRepository.js
@@ -1,0 +1,11 @@
+export default class PaymentsRepository {
+
+    /**
+     *
+     * @returns {*}
+     */
+    getSubmitButton() {
+        return cy.get('#confirmPaymentForm > .btn-primary');
+    }
+
+}
