+++ tests/Cypress/cypress/support/repositories/storefront/navigation/NavigationRepository.js
@@ -1,0 +1,11 @@
+export default class NavigationRepository {
+
+    /**
+     *
+     * @returns {*}
+     */
+    getHomeMenuItem() {
+        return cy.get('.home-link > .main-navigation-link-text');
+    }
+
+}
