+++ tests/Cypress/cypress/support/repositories/storefront/products/ListingRepository.js
@@ -1,0 +1,11 @@
+export default class ListingRepository {
+
+    /**
+     *
+     * @returns {*}
+     */
+    getFirstProduct() {
+        return cy.get(':nth-child(1) > .card > .card-body');
+    }
+
+}
