+++ tests/Cypress/cypress/support/repositories/storefront/products/PDPRepository.js
@@ -1,0 +1,11 @@
+export default class PDPRepository {
+
+    /**
+     *
+     * @returns {*}
+     */
+    getAddToCartButton() {
+        return cy.get('.buy-widget-container > .col-8 > .btn');
+    }
+
+}
